- Timing now synchronizes properly on Linux.  Before, it just went at full speed all the time.  In Time::SetTotalMilliseconds, it was dividing totalMilliseconds by 1000 after adding the seconds.  This was wrong.  It now works, but I had trouble with floating point precision issues.  If you remove the temporaries (numSeconds and numNanoseconds), the function will add a second sometimes. refs #98

+++ Source/LinuxPlatform/Thread.cpp
@@ -115,7 +115,7 @@ void Thread::Sleep(unsigned int milliseconds)
 {
     Time startTime = Time::Now();
 
-    long int elapsedMilliseconds = 0;
+    float elapsedMilliseconds = 0.f;
 
     do
     {
@@ -123,6 +123,6 @@ void Thread::Sleep(unsigned int milliseconds)
 
         elapsedMilliseconds = Time::Now().GetTotalMilliseconds() - startTime.GetTotalMilliseconds();
 
-    }   while(elapsedMilliseconds < milliseconds);
+    }   while((unsigned int)elapsedMilliseconds < milliseconds);
 }
